+++ src/templates/login.js
@@ -1,25 +1,24 @@
-import React from 'react';
+import React from "react";
+import LoginForm from "../components/LoginForm";
+import "../sass/custom.scss";
 
-import {Layout} from '../components/index';
-import { logIn} from '../services/auth';
+const Login = () => {
+  return (
+    <section className="custom-container">
+      <div className="custom-container__left">
+        <div className="content">
+          <h2>Hello</h2>
+          <p>Sign in to your account to continue</p>
+        </div>
+      </div>
+      <div className="custom-container__right">
+        <div className="content">
+          <h3 className="title">Sign in</h3>
+          <LoginForm />
+        </div>
+      </div>
+    </section>
+  );
+};
 
-
-export default class Post extends React.Component {
-    componentDidMount(){
-       
-    }
-
-    handleLogin = () => {
-        logIn('t@t.com', '123')
-    }
-
-    render() {
-        return (
-            <Layout {...this.props}>
-              <button onClick={this.handleLogin}> 
-                  signup
-              </button>
-            </Layout>
-        );
-    }
-}
+export default Login;
